QQ: move rescale and add compose with generators
