+++ fmpq_mpoly/test/t-compose_fmpq_mpoly.c
@@ -16,11 +16,158 @@
 int
 main(void)
 {
-    slong i, v;
+    slong i, j, v;
     FLINT_TEST_INIT(state);
 
     flint_printf("compose_fmpq_mpoly....");
     fflush(stdout);
+
+    {
+        fmpq_mpoly_t A, A1, A2, B;
+        fmpq_mpoly_struct * Cp[3];
+        fmpq_mpoly_struct C[3];
+        fmpq_mpoly_ctx_t ctxAC, ctxB;
+
+        fmpq_mpoly_ctx_init(ctxB, 3, ORD_LEX);
+        fmpq_mpoly_ctx_init(ctxAC, 2, ORD_LEX);
+
+        fmpq_mpoly_init(B, ctxB);
+        fmpq_mpoly_init(A, ctxAC);
+        fmpq_mpoly_init(A1, ctxAC);
+        fmpq_mpoly_init(A2, ctxAC);
+        for (i = 0; i < 3; i++)
+        {
+            Cp[i] = C + i;
+            fmpq_mpoly_init(C + i, ctxAC);
+        }
+
+        fmpq_mpoly_set_str_pretty(B,
+                "1 + x1*x2^2 + x2^9999999999999999999999999*x3^9", NULL, ctxB);
+
+        fmpq_mpoly_set_str_pretty(C + 0, "x1 + x2", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 1, "x1 - x2", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 2, "1", NULL, ctxAC);
+        if (fmpq_mpoly_compose_fmpq_mpoly(A, B, Cp, ctxB, ctxAC))
+        {
+            printf("FAIL\n");
+            flint_printf("Check non-example 1\n", i);
+            flint_abort();
+        }
+
+        fmpq_mpoly_set_str_pretty(C + 0, "x1", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 1, "2*x2", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 2, "1", NULL, ctxAC);
+        if (fmpq_mpoly_compose_fmpq_mpoly(A, B, Cp, ctxB, ctxAC))
+        {
+            printf("FAIL\n");
+            flint_printf("Check non-example 2\n", i);
+            flint_abort();
+        }
+
+        fmpq_mpoly_set_str_pretty(C + 0, "2*x1", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 1, "x2", NULL, ctxAC);
+        fmpq_mpoly_set_str_pretty(C + 2, "1", NULL, ctxAC);
+        if (!fmpq_mpoly_compose_fmpq_mpoly(A, B, Cp, ctxB, ctxAC))
+        {
+            printf("FAIL\n");
+            flint_printf("Check example 3\n", i);
+            flint_abort();
+        }
+
+        fmpq_mpoly_clear(B, ctxB);
+        fmpq_mpoly_clear(A, ctxAC);
+        fmpq_mpoly_clear(A1, ctxAC);
+        fmpq_mpoly_clear(A2, ctxAC);
+        for (i = 0; i < 3; i++)
+            fmpq_mpoly_clear(C + i, ctxAC);
+
+        fmpq_mpoly_ctx_clear(ctxB);
+        fmpq_mpoly_ctx_clear(ctxAC);
+    }
+
+    /* check composition with generators */
+    for (i = 0; i < 50*flint_test_multiplier(); i++)
+    {
+        fmpq_mpoly_ctx_t ctxB, ctxAC;
+        fmpq_mpoly_t B, A, A1;
+        fmpq_mpoly_struct ** C;
+        slong * c;
+        slong nvarsB, nvarsAC;
+        slong len;
+        flint_bitcnt_t exp_bits, coeff_bits;
+
+        fmpq_mpoly_ctx_init_rand(ctxB, state, 20);
+        fmpq_mpoly_ctx_init_rand(ctxAC, state, 20);
+
+        nvarsB = ctxB->zctx->minfo->nvars;
+        nvarsAC = ctxAC->zctx->minfo->nvars;
+
+        fmpq_mpoly_init(B, ctxB);
+        fmpq_mpoly_init(A, ctxAC);
+        fmpq_mpoly_init(A1, ctxAC);
+
+        c = (slong *) flint_malloc(nvarsB*sizeof(slong));
+        C = (fmpq_mpoly_struct **) flint_malloc(nvarsB * sizeof(fmpq_mpoly_struct *));
+        for (v = 0; v < nvarsB; v++)
+        {
+            C[v] = (fmpq_mpoly_struct *) flint_malloc(sizeof(fmpq_mpoly_struct));
+            fmpq_mpoly_init(C[v], ctxAC);
+        }
+
+        for (j = 0; j < 4; j++)
+        {
+            len = n_randint(state, 200);
+            exp_bits = n_randint(state, 100) + 1;
+            coeff_bits = n_randint(state, 100) + 1;
+
+            for (v = 0; v < nvarsB; v++)
+            {
+                c[v] = n_randint(state, nvarsAC + 2) - 2;
+                if (c[v] >= 0)
+                    fmpq_mpoly_gen(C[v], c[v], ctxAC);
+                else
+                    fmpq_mpoly_zero(C[v], ctxAC);
+            }
+
+            fmpq_mpoly_randtest_bits(B, state, len, coeff_bits, exp_bits, ctxB);
+
+            fmpq_mpoly_compose_fmpq_mpoly_gen(A, B, c, ctxB, ctxAC);
+
+            if (!fmpq_mpoly_compose_fmpq_mpoly(A1, B, C, ctxB, ctxAC))
+            {
+                printf("FAIL\n");
+                flint_printf("Check composition success with generators\n"
+                                                     "i: %wd, j: %wd\n", i, j);
+                flint_abort();
+            }
+
+            if (!fmpq_mpoly_equal(A, A1, ctxAC))
+            {
+                printf("FAIL\n");
+                flint_printf("Check composition with generators\n"
+                                                     "i: %wd, j: %wd\n", i, j);
+                flint_abort();
+            }
+
+            fmpq_mpoly_assert_canonical(A, ctxAC);
+            fmpq_mpoly_assert_canonical(A1, ctxAC);
+        }
+
+        for (v = 0; v < nvarsB; v++)
+        {
+            fmpq_mpoly_clear(C[v], ctxAC);
+            flint_free(C[v]);
+        }
+        flint_free(C);
+        flint_free(c);
+
+        fmpq_mpoly_clear(B, ctxB);
+        fmpq_mpoly_clear(A, ctxAC);
+        fmpq_mpoly_clear(A1, ctxAC);
+
+        fmpq_mpoly_ctx_clear(ctxB);
+        fmpq_mpoly_ctx_clear(ctxAC);
+    }
 
     /* Check composition with identity */
     for (i = 0; i < 50*flint_test_multiplier(); i++)
